Hurray! We have successfully uploaded an blog.
Today, we have done work on submit function on add blog page in which we first check all the fields are empty or not with sub-headings inputs and also heading image. Then, We have uploaded all the images in the firebase storage and all the other data including images (url,path) in Realtime database. And we have also given an option to user for multiple images. They can be able to upload that images according to their will.