Change on AddBlog.js page and deletion of a blog.
Today, we have successfully deleted a blog in which we first check optional images are present or not. If present, then we deleted them and after that we deleted heading image from Firebase Storage. And at last, we deleted the object which contains all other information from Firebase Realtime database. And we randomly created random path for all the images using Date function.