Sorting and Slicing
Yesterday, we have done the sorting of all the blogs from firebase Realtime database according to the date when they are uploaded and from that we fetching first 10 to 15 objects using slice function on array. And used this JSON data to show multiple latest images on the bottom of the home page and done the work of trending now section on home page by just 3 blogs.