Started work on My Account page for user details
At that day, we have just started the work on My Account page in which we first of all fetched user details from Firebase Realtime Database using user token stored in local storage and check whether user is authenticated or not. And then we use conditional rendering whether user profile image is present or not. If it is not present then we have uploaded that image is firebase storage and updating user info in firebase Realtime Database and successfully stored its Image URL and FullPath